Remove tbtc.ts file
Remove `tbtc.ts` from threshold lib and keep everything in `tbtc/index.ts` file.
